Plando: Move huge constants into their own file.
Also separate out Koot Favor Items, Coins and Rewards.
