+++ app/src/app/pages/plando-page/plando-constants.ts
@@ -1,0 +1,1214 @@
+export enum CheckType {
+    NORMAL,
+    SHOP,
+    HIDDEN_PANEL,
+    COIN_BLOCK,
+    MULTICOIN_BLOCK,
+    SUPER_BLOCK,
+    OVERWORLD_COIN,
+    TRADE_EVENT,
+    FOLIAGE,
+    KOOT_FAVOR_ITEM,
+    KOOT_FAVOR_COIN,
+    KOOT_FAVOR_REWARD,
+    LETTER_REWARD
+}
+
+type Check = {
+    name: string;
+    type: CheckType;
+}
+
+export type Location = {
+    name: string;
+    checks: Array<Check>;
+}
+
+export const LOCATIONS_LIST: Array<Location> = [
+    {
+        name: "Goomba Region", checks: [
+            { name: "Forest Clearing - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Goomba Village - Bush Bottom Right", type: CheckType.FOLIAGE },
+            { name: "Goomba Village - Goompa Koopa Koot Favor", type: CheckType.KOOT_FAVOR_ITEM },
+            { name: "Goomba Village - Goompa Gift", type: CheckType.NORMAL },
+            { name: "Goomba Village - Goombaria Dolly Reward", type: CheckType.NORMAL },
+            { name: "Goomba Village - Goompa Letter Reward", type: CheckType.LETTER_REWARD },
+            { name: "Goomba Village - Goompapa Letter Reward 1", type: CheckType.LETTER_REWARD },
+            { name: "Goomba Village - Goompapa Letter Reward 2", type: CheckType.LETTER_REWARD },
+            { name: "Goomba Village - On The Balcony", type: CheckType.NORMAL },
+            { name: "Goomba Village - Goombario Partner", type: CheckType.NORMAL },
+            { name: "Goomba Village - Goomnut Tree", type: CheckType.FOLIAGE },
+            { name: "Behind the Village - On Ledge", type: CheckType.NORMAL },
+            { name: "Behind the Village - In Tree", type: CheckType.FOLIAGE },
+            { name: "Bottom of the Cliff - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Bottom of the Cliff - Above Stone Block", type: CheckType.NORMAL },
+            { name: "Bottom of the Cliff - Floating Coin 1", type: CheckType.OVERWORLD_COIN },
+            { name: "Bottom of the Cliff - Floating Coin 2", type: CheckType.OVERWORLD_COIN },
+            { name: "Bottom of the Cliff - Floating Coin 3", type: CheckType.OVERWORLD_COIN },
+            { name: "Bottom of the Cliff - Floating Coin 4", type: CheckType.OVERWORLD_COIN },
+            { name: "Bottom of the Cliff - Upper Ledge", type: CheckType.NORMAL },
+            { name: "Bottom of the Cliff - In Tree", type: CheckType.FOLIAGE },
+            { name: "Bottom of the Cliff - Block On Ground", type: CheckType.COIN_BLOCK },
+            { name: "Jr. Troopas Playground - Bush Right", type: CheckType.FOLIAGE },
+            { name: "Jr. Troopas Playground - Bush Bottom Right", type: CheckType.FOLIAGE },
+            { name: "Jr. Troopas Playground - Bush Top 1", type: CheckType.FOLIAGE },
+            { name: "Jr. Troopas Playground - Bush Top 2", type: CheckType.FOLIAGE },
+            { name: "Jr. Troopas Playground - Bush Center", type: CheckType.FOLIAGE },
+            { name: "Jr. Troopas Playground - Bush Top Left", type: CheckType.FOLIAGE },
+            { name: "Jr. Troopas Playground - In Hammer Bush", type: CheckType.FOLIAGE },
+            { name: "Jr. Troopas Playground - In MultiCoinBlock", type: CheckType.MULTICOIN_BLOCK },
+            { name: "Jr. Troopas Playground - In Tree Left", type: CheckType.FOLIAGE },
+            { name: "Jr. Troopas Playground - In Tree Top", type: CheckType.FOLIAGE },
+            { name: "Jr. Troopas Playground - In Tree Right", type: CheckType.FOLIAGE },
+            { name: "Goomba Road 1 - Yellow Block Left", type: CheckType.COIN_BLOCK },
+            { name: "Goomba Road 1 - Yellow Block Right", type: CheckType.COIN_BLOCK },
+            { name: "Goomba Road 2 - On the Sign", type: CheckType.NORMAL },
+            { name: "Goomba Road 2 - Red Block", type: CheckType.NORMAL },
+            { name: "Goomba Kings Castle - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Goomba Kings Castle - In Tree Left Of Fortress", type: CheckType.FOLIAGE },
+            { name: "Goomba Kings Castle - In Tree Right Of Cliff", type: CheckType.FOLIAGE },
+            { name: "Goomba Kings Castle - Hidden Yellow Block", type: CheckType.NORMAL },
+            { name: "Toad Town Entrance - Chest On Roof", type: CheckType.NORMAL },
+            { name: "Toad Town Entrance - Yellow Block", type: CheckType.NORMAL },
+            { name: "Marios House - Luigi Koopa Koot Favor", type: CheckType.KOOT_FAVOR_ITEM },
+        ]
+    },
+    {
+        name: "Toad Town", checks: [
+            { name: "Gate District - Dojo: Chan", type: CheckType.NORMAL },
+            { name: "Gate District - Dojo: Lee", type: CheckType.NORMAL },
+            { name: "Gate District - Dojo: Master 1", type: CheckType.NORMAL },
+            { name: "Gate District - Dojo: Master 2", type: CheckType.NORMAL },
+            { name: "Gate District - Dojo: Master 3", type: CheckType.NORMAL },
+            { name: "Gate District - Russ T. Dictionary Reward", type: CheckType.NORMAL },
+            { name: "Gate District - Russ T. Letter Reward", type: CheckType.LETTER_REWARD },
+            { name: "Gate District - Miss T. Letter Reward", type: CheckType.LETTER_REWARD },
+            { name: "Gate District - Radio Trade Event 1 Reward", type: CheckType.TRADE_EVENT },
+            { name: "Gate District - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Gate District - Sushie Island", type: CheckType.NORMAL },
+            { name: "Gate District - Shop Item 1", type: CheckType.SHOP },
+            { name: "Gate District - Shop Item 2", type: CheckType.SHOP },
+            { name: "Gate District - Shop Item 3", type: CheckType.SHOP },
+            { name: "Gate District - Shop Item 4", type: CheckType.SHOP },
+            { name: "Gate District - Shop Item 5", type: CheckType.SHOP },
+            { name: "Gate District - Shop Item 6", type: CheckType.SHOP },
+            { name: "Plaza District - Rowf Calculator Reward", type: CheckType.NORMAL },
+            { name: "Plaza District - Postmaster MailBag Reward", type: CheckType.NORMAL },
+            { name: "Plaza District - Merlon Letter Reward", type: CheckType.LETTER_REWARD },
+            { name: "Plaza District - Minh T. Letter Reward", type: CheckType.LETTER_REWARD },
+            { name: "Plaza District - Merlon House Stomp x3", type: CheckType.NORMAL },
+            { name: "Plaza District - Rowfs Shop Set 1 - 1", type: CheckType.SHOP },
+            { name: "Plaza District - Rowfs Shop Set 1 - 2", type: CheckType.SHOP },
+            { name: "Plaza District - Rowfs Shop Set 1 - 3", type: CheckType.SHOP },
+            { name: "Plaza District - Rowfs Shop Set 1 - 4", type: CheckType.SHOP },
+            { name: "Plaza District - Rowfs Shop Set 2 - 1", type: CheckType.SHOP },
+            { name: "Plaza District - Rowfs Shop Set 2 - 2", type: CheckType.SHOP },
+            { name: "Plaza District - Rowfs Shop Set 2 - 3", type: CheckType.SHOP },
+            { name: "Plaza District - Rowfs Shop Set 3 - 1", type: CheckType.SHOP },
+            { name: "Plaza District - Rowfs Shop Set 3 - 2", type: CheckType.SHOP },
+            { name: "Plaza District - Rowfs Shop Set 3 - 3", type: CheckType.SHOP },
+            { name: "Plaza District - Rowfs Shop Set 4 - 1", type: CheckType.SHOP },
+            { name: "Plaza District - Rowfs Shop Set 4 - 2", type: CheckType.SHOP },
+            { name: "Plaza District - Rowfs Shop Set 4 - 3", type: CheckType.SHOP },
+            { name: "Plaza District - Rowfs Shop Set 5 - 1", type: CheckType.SHOP },
+            { name: "Plaza District - Rowfs Shop Set 5 - 2", type: CheckType.SHOP },
+            { name: "Plaza District - Rowfs Shop Set 5 - 3", type: CheckType.SHOP },
+            { name: "Plaza District - In Tree", type: CheckType.FOLIAGE },
+            { name: "Southern District - Bub-ulb Gift", type: CheckType.NORMAL },
+            { name: "Southern District - Tayce T. Frying Pan Reward", type: CheckType.NORMAL },
+            { name: "Southern District - Fice T. Letter Reward", type: CheckType.LETTER_REWARD },
+            { name: "Southern District - Fice T. Forest Pass", type: CheckType.NORMAL },
+            { name: "Southern District - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Southern District - Inside Blue House", type: CheckType.NORMAL },
+            { name: "Station District - Dane T. Letter Reward 1", type: CheckType.LETTER_REWARD },
+            { name: "Station District - Dane T. Letter Reward 2", type: CheckType.LETTER_REWARD },
+            { name: "Station District - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Residental District - Storeroom Item 1", type: CheckType.NORMAL },
+            { name: "Residental District - Storeroom Item 2", type: CheckType.NORMAL },
+            { name: "Residental District - Storeroom Item 3", type: CheckType.NORMAL },
+            { name: "Residental District - Storeroom Item 4", type: CheckType.NORMAL },
+            { name: "Residental District - Shop Item 1", type: CheckType.SHOP },
+            { name: "Residental District - Shop Item 2", type: CheckType.SHOP },
+            { name: "Residental District - Shop Item 3", type: CheckType.SHOP },
+            { name: "Residental District - Shop Item 4", type: CheckType.SHOP },
+            { name: "Residental District - Shop Item 5", type: CheckType.SHOP },
+            { name: "Residental District - Shop Item 6", type: CheckType.SHOP },
+            { name: "Port District - Poet Gift", type: CheckType.NORMAL },
+            { name: "Port District - Poet Melody Reward", type: CheckType.NORMAL },
+            { name: "Port District - Fishmael Letter Reward", type: CheckType.LETTER_REWARD },
+            { name: "Port District - Radio Trade Event 3 Reward", type: CheckType.TRADE_EVENT },
+            { name: "Port District - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Port District - In MultiCoinBlock", type: CheckType.MULTICOIN_BLOCK },
+            { name: "Ruined Castle Grounds - Muss T. Letter Reward", type: CheckType.LETTER_REWARD },
+        ]
+    },
+    {
+        name: "Toad Town Tunnels", checks: [
+            { name: "Hall to Blooper 1 (B1) - Hidden Block", type: CheckType.NORMAL },
+            { name: "Hall to Blooper 1 (B1) - In MultiCoinBlock", type: CheckType.MULTICOIN_BLOCK },
+            { name: "Blooper Boss 1 (B1) - Blooper Fight Reward", type: CheckType.NORMAL },
+            { name: "Short Elevator Room (B1) - Yellow Block Center", type: CheckType.NORMAL },
+            { name: "Short Elevator Room (B1) - Yellow Block Left", type: CheckType.COIN_BLOCK },
+            { name: "Short Elevator Room (B1) - Yellow Block Right", type: CheckType.COIN_BLOCK },
+            { name: "Spring Room (B2) - Chest On Ledge", type: CheckType.NORMAL },
+            { name: "Elevator Attic Room (B2) - On Parakarry Ledge", type: CheckType.NORMAL },
+            { name: "Elevator Attic Room (B2) - In SuperBlock", type: CheckType.SUPER_BLOCK },
+            { name: "Metal Block Room (B3) - In SuperBlock", type: CheckType.SUPER_BLOCK },
+            { name: "Blue Pushblock Room (B2) - Hidden Block Left", type: CheckType.COIN_BLOCK },
+            { name: "Blue Pushblock Room (B2) - Hidden Block Center", type: CheckType.COIN_BLOCK },
+            { name: "Blue Pushblock Room (B2) - Hidden Block Right", type: CheckType.COIN_BLOCK },
+            { name: "Blue Pushblock Room (B2) - In SuperBlock", type: CheckType.SUPER_BLOCK },
+            { name: "Room with Spikes (B2) - Yellow Block", type: CheckType.NORMAL },
+            { name: "Winding Path (Spiny Room) - Hidden Block Center", type: CheckType.NORMAL },
+            { name: "Winding Path (Spiny Room) - Hidden Block Right", type: CheckType.NORMAL },
+            { name: "Winding Path (Spiny Room) - Hidden Block Left", type: CheckType.NORMAL },
+            { name: "Winding Path (Spiny Room) - Yellow Block", type: CheckType.COIN_BLOCK },
+            { name: "Hall to Ultra Boots (B3) - Hidden Block", type: CheckType.NORMAL },
+            { name: "Hall to Ultra Boots (B3) - Yellow Block Left", type: CheckType.COIN_BLOCK },
+            { name: "Hall to Ultra Boots (B3) - Yellow Block Right", type: CheckType.COIN_BLOCK },
+            { name: "Ultra Boots Room (B3) - In Big Chest", type: CheckType.NORMAL },
+            { name: "Bridge to Shiver City (B2) - Yellow Block 1", type: CheckType.COIN_BLOCK },
+            { name: "Bridge to Shiver City (B2) - Yellow Block 2", type: CheckType.COIN_BLOCK },
+            { name: "Bridge to Shiver City (B2) - Yellow Block 3", type: CheckType.COIN_BLOCK },
+            { name: "Bridge to Shiver City (B2) - Yellow Block 4", type: CheckType.COIN_BLOCK },
+            { name: "Bridge to Shiver City (B2) - Yellow Block 5", type: CheckType.COIN_BLOCK },
+            { name: "Frozen Room (B3) - In SuperBlock", type: CheckType.SUPER_BLOCK },
+            { name: "Rip Cheatos Home (B3) - Rip Cheato Offer 1", type: CheckType.NORMAL },
+            { name: "Rip Cheatos Home (B3) - Rip Cheato Offer 2", type: CheckType.NORMAL },
+            { name: "Rip Cheatos Home (B3) - Rip Cheato Offer 3", type: CheckType.NORMAL },
+            { name: "Rip Cheatos Home (B3) - Rip Cheato Offer 4", type: CheckType.NORMAL },
+            { name: "Rip Cheatos Home (B3) - Rip Cheato Offer 5", type: CheckType.NORMAL },
+            { name: "Rip Cheatos Home (B3) - Rip Cheato Offer 6", type: CheckType.NORMAL },
+            { name: "Rip Cheatos Home (B3) - Rip Cheato Offer 7", type: CheckType.NORMAL },
+            { name: "Rip Cheatos Home (B3) - Rip Cheato Offer 8", type: CheckType.NORMAL },
+            { name: "Rip Cheatos Home (B3) - Rip Cheato Offer 9", type: CheckType.NORMAL },
+            { name: "Rip Cheatos Home (B3) - Rip Cheato Offer 10", type: CheckType.NORMAL },
+            { name: "Rip Cheatos Home (B3) - Rip Cheato Offer 11", type: CheckType.NORMAL },
+            { name: "Under the Toad Town Pond - In SuperBlock", type: CheckType.SUPER_BLOCK },
+        ]
+    },
+    {
+        name: "Shooting Star Summit", checks: [
+            { name: "Shooting Star Path - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Merluvlees House - Merluvlee Koopa Koot Favor", type: CheckType.KOOT_FAVOR_ITEM },
+            { name: "Merluvlees House - Merlow Letter Reward", type: CheckType.LETTER_REWARD },
+            { name: "Merluvlees House - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Merluvlees House - Merlows Badges 1", type: CheckType.NORMAL },
+            { name: "Merluvlees House - Merlows Badges 2", type: CheckType.NORMAL },
+            { name: "Merluvlees House - Merlows Badges 3", type: CheckType.NORMAL },
+            { name: "Merluvlees House - Merlows Badges 4", type: CheckType.NORMAL },
+            { name: "Merluvlees House - Merlows Badges 5", type: CheckType.NORMAL },
+            { name: "Merluvlees House - Merlows Badges 6", type: CheckType.NORMAL },
+            { name: "Merluvlees House - Merlows Badges 7", type: CheckType.NORMAL },
+            { name: "Merluvlees House - Merlows Badges 8", type: CheckType.NORMAL },
+            { name: "Merluvlees House - Merlows Badges 9", type: CheckType.NORMAL },
+            { name: "Merluvlees House - Merlows Badges 10", type: CheckType.NORMAL },
+            { name: "Merluvlees House - Merlows Badges 11", type: CheckType.NORMAL },
+            { name: "Merluvlees House - Merlows Badges 12", type: CheckType.NORMAL },
+            { name: "Merluvlees House - Merlows Badges 13", type: CheckType.NORMAL },
+            { name: "Merluvlees House - Merlows Badges 14", type: CheckType.NORMAL },
+            { name: "Merluvlees House - Merlows Badges 15", type: CheckType.NORMAL },
+            { name: "Merluvlees House - Merlows Rewards 1", type: CheckType.NORMAL },
+            { name: "Merluvlees House - Merlows Rewards 2", type: CheckType.NORMAL },
+            { name: "Merluvlees House - Merlows Rewards 3", type: CheckType.NORMAL },
+            { name: "Merluvlees House - Merlows Rewards 4", type: CheckType.NORMAL },
+            { name: "Merluvlees House - Merlows Rewards 5", type: CheckType.NORMAL },
+            { name: "Merluvlees House - Merlows Rewards 6", type: CheckType.NORMAL },
+            { name: "Shooting Star Summit - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Shooting Star Summit - Behind The Summit", type: CheckType.NORMAL },
+            { name: "Star Haven - Shop Item 1", type: CheckType.SHOP },
+            { name: "Star Haven - Shop Item 2", type: CheckType.SHOP },
+            { name: "Star Haven - Shop Item 3", type: CheckType.SHOP },
+            { name: "Star Haven - Shop Item 4", type: CheckType.SHOP },
+            { name: "Star Haven - Shop Item 5", type: CheckType.SHOP },
+            { name: "Star Haven - Shop Item 6", type: CheckType.SHOP },
+            { name: "Star Sanctuary - Gift of the Stars", type: CheckType.NORMAL },
+        ]
+    },
+    {
+        name: "Koopa Region", checks: [
+            { name: "Pleasant Path Entry - Red Block Center", type: CheckType.NORMAL },
+            { name: "Pleasant Path Entry - Yellow Block Left", type: CheckType.COIN_BLOCK },
+            { name: "Pleasant Path Entry - Yellow Block Right", type: CheckType.NORMAL },
+            { name: "Pleasant Path Bridge - Kooper Island", type: CheckType.NORMAL },
+            { name: "Pleasant Path Bridge - Behind Fence", type: CheckType.NORMAL },
+            { name: "Pleasant Path Bridge - In MultiCoinBlock", type: CheckType.MULTICOIN_BLOCK },
+            { name: "Pleasant Path Bridge - Yellow Block", type: CheckType.NORMAL },
+            { name: "Pleasant Crossroads - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Pleasant Crossroads - Behind Peg", type: CheckType.NORMAL },
+            { name: "Pleasant Crossroads - Brick Block Puzzle", type: CheckType.NORMAL },
+            { name: "Path to Fortress 1 - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Path to Fortress 1 - Hidden Block", type: CheckType.NORMAL },
+            { name: "Path to Fortress 1 - X On Ground 1", type: CheckType.OVERWORLD_COIN },
+            { name: "Path to Fortress 1 - X On Ground 2", type: CheckType.OVERWORLD_COIN },
+            { name: "Path to Fortress 1 - X On Ground 3", type: CheckType.OVERWORLD_COIN },
+            { name: "Path to Fortress 1 - X On Ground 4", type: CheckType.OVERWORLD_COIN },
+            { name: "Path to Fortress 1 - X On Ground 5", type: CheckType.OVERWORLD_COIN },
+            { name: "Path to Fortress 1 - On Brick Block", type: CheckType.NORMAL },
+            { name: "Path to Fortress 2 - In Tree", type: CheckType.FOLIAGE },
+            { name: "Koopa Village 1 - Bush Far Left", type: CheckType.FOLIAGE },
+            { name: "Koopa Village 1 - Bush Left Front", type: CheckType.FOLIAGE },
+            { name: "Koopa Village 1 - Bush Infront Of Tree", type: CheckType.FOLIAGE },
+            { name: "Koopa Village 1 - Bush Second From Right", type: CheckType.FOLIAGE },
+            { name: "Koopa Village 1 - Bush Second From Left (Koopa Koot Favor)", type: CheckType.KOOT_FAVOR_ITEM },
+            { name: "Koopa Village 1 - Bush Far Right (Koopa Koot Favor)", type: CheckType.KOOT_FAVOR_ITEM },
+            { name: "Koopa Village 1 - Mort T. Letter Reward", type: CheckType.LETTER_REWARD },
+            { name: "Koopa Village 1 - Koover Letter Reward 1", type: CheckType.LETTER_REWARD },
+            { name: "Koopa Village 1 - Koover Letter Reward 2", type: CheckType.LETTER_REWARD },
+            { name: "Koopa Village 1 - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Koopa Village 1 - Shop Item 1", type: CheckType.SHOP },
+            { name: "Koopa Village 1 - Shop Item 2", type: CheckType.SHOP },
+            { name: "Koopa Village 1 - Shop Item 3", type: CheckType.SHOP },
+            { name: "Koopa Village 1 - Shop Item 4", type: CheckType.SHOP },
+            { name: "Koopa Village 1 - Shop Item 5", type: CheckType.SHOP },
+            { name: "Koopa Village 1 - Shop Item 6", type: CheckType.SHOP },
+            { name: "Koopa Village 2 - Bush Far Left", type: CheckType.FOLIAGE },
+            { name: "Koopa Village 2 - Bush Far Right", type: CheckType.FOLIAGE },
+            { name: "Koopa Village 2 - Kolorados Wife (Koopa Koot Favor)", type: CheckType.KOOT_FAVOR_ITEM },
+            { name: "Koopa Village 2 - Favor 2 Reward 2 (SleepySheep)", type: CheckType.KOOT_FAVOR_REWARD },
+            { name: "Koopa Village 2 - Favor 10 Reward 2 (LifeShroom)", type: CheckType.KOOT_FAVOR_REWARD },
+            { name: "Koopa Village 2 - Kolorado Artifact Reward", type: CheckType.NORMAL },
+            { name: "Koopa Village 2 - Kolorado Letter Reward", type: CheckType.LETTER_REWARD },
+            { name: "Koopa Village 2 - Push Block Puzzle", type: CheckType.NORMAL },
+            { name: "Koopa Village 2 - Favor 1 Reward (KoopaLegends)", type: CheckType.KOOT_FAVOR_COIN },
+            { name: "Koopa Village 2 - Favor 2 Reward 1 (SleepySheep)", type: CheckType.KOOT_FAVOR_COIN },
+            { name: "Koopa Village 2 - Favor 3 Reward (Tape)", type: CheckType.KOOT_FAVOR_COIN },
+            { name: "Koopa Village 2 - Favor 4 Reward (KoopaTea)", type: CheckType.KOOT_FAVOR_REWARD },
+            { name: "Koopa Village 2 - Favor 5 Reward (LuigiAutograph)", type: CheckType.KOOT_FAVOR_COIN },
+            { name: "Koopa Village 2 - Favor 6 Reward (Wallet)", type: CheckType.KOOT_FAVOR_COIN },
+            { name: "Koopa Village 2 - Favor 7 Reward (TastyTonic)", type: CheckType.KOOT_FAVOR_COIN },
+            { name: "Koopa Village 2 - Favor 8 Reward (MerluvleeAutograph)", type: CheckType.KOOT_FAVOR_REWARD },
+            { name: "Koopa Village 2 - Favor 9 Reward (News)", type: CheckType.KOOT_FAVOR_COIN },
+            { name: "Koopa Village 2 - Favor 10 Reward 1 (LifeShroom)", type: CheckType.KOOT_FAVOR_COIN },
+            { name: "Koopa Village 2 - Favor 11 Reward (NuttyCake)", type: CheckType.KOOT_FAVOR_COIN },
+            { name: "Koopa Village 2 - Favor 12 Reward (Bob-ombs)", type: CheckType.KOOT_FAVOR_REWARD },
+            { name: "Koopa Village 2 - Favor 13 Reward (OldPhoto)", type: CheckType.KOOT_FAVOR_COIN },
+            { name: "Koopa Village 2 - Favor 14 Reward (Koopasta)", type: CheckType.KOOT_FAVOR_COIN },
+            { name: "Koopa Village 2 - Favor 15 Reward (Glasses)", type: CheckType.KOOT_FAVOR_COIN },
+            { name: "Koopa Village 2 - Favor 16 Reward (Lime)", type: CheckType.KOOT_FAVOR_REWARD },
+            { name: "Koopa Village 2 - Favor 17 Reward (KookyCookie)", type: CheckType.KOOT_FAVOR_COIN },
+            { name: "Koopa Village 2 - Favor 18 Reward (Package)", type: CheckType.KOOT_FAVOR_COIN },
+            { name: "Koopa Village 2 - Favor 19 Reward (Coconut)", type: CheckType.KOOT_FAVOR_COIN },
+            { name: "Koopa Village 2 - Favor 20 Reward (RedJar)", type: CheckType.KOOT_FAVOR_REWARD },
+            { name: "Koopa Village 2 - Kooper Partner", type: CheckType.NORMAL },
+            { name: "Behind Koopa Village - On Stump", type: CheckType.NORMAL },
+            { name: "Fuzzy Forest - Fuzzy Battle Reward", type: CheckType.NORMAL },
+        ]
+    },
+    {
+        name: "Koopa Bros Fortress", checks: [
+            { name: "Fortress Exterior - Chest Behind Fortress", type: CheckType.NORMAL },
+            { name: "Fortress Exterior - Chest On Ledge", type: CheckType.NORMAL },
+            { name: "Left Tower - Top Of Tower", type: CheckType.NORMAL },
+            { name: "Left Tower - Koopa Troopa Reward", type: CheckType.NORMAL },
+            { name: "Central Hall - Left Cell", type: CheckType.NORMAL },
+            { name: "Central Hall - Right Cell", type: CheckType.NORMAL },
+            { name: "Central Hall - Center Cell", type: CheckType.NORMAL },
+            { name: "Jail - Bombette Partner", type: CheckType.NORMAL },
+            { name: "Dungeon Fire Room - On The Ground", type: CheckType.NORMAL },
+            { name: "Battlement - Block Behind Rock", type: CheckType.NORMAL },
+        ]
+    },
+    {
+        name: "Mt Rugged", checks: [
+            { name: "Train Station - Bush 1", type: CheckType.FOLIAGE },
+            { name: "Train Station - Bush 2", type: CheckType.FOLIAGE },
+            { name: "Train Station - Bush 3", type: CheckType.FOLIAGE },
+            { name: "Train Station - Bush Top", type: CheckType.FOLIAGE },
+            { name: "Train Station - Parakarry Partner", type: CheckType.NORMAL },
+            { name: "Train Station - In SuperBlock", type: CheckType.SUPER_BLOCK },
+            { name: "Mt Rugged 1 - On Slide 1", type: CheckType.OVERWORLD_COIN },
+            { name: "Mt Rugged 1 - On Slide 2", type: CheckType.OVERWORLD_COIN },
+            { name: "Mt Rugged 1 - On Slide 3", type: CheckType.OVERWORLD_COIN },
+            { name: "Mt Rugged 1 - Hurting Whacka", type: CheckType.NORMAL },
+            { name: "Mt Rugged 1 - Yellow Block", type: CheckType.NORMAL },
+            { name: "Mt Rugged 2 - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Mt Rugged 2 - Parakarry Ledge", type: CheckType.NORMAL },
+            { name: "Mt Rugged 2 - Kooper Ledge", type: CheckType.NORMAL },
+            { name: "Mt Rugged 3 - Bub-ulb Gift", type: CheckType.NORMAL },
+            { name: "Mt Rugged 3 - On Scaffolding", type: CheckType.NORMAL },
+            { name: "Mt Rugged 4 - Hidden Cave Chest", type: CheckType.NORMAL },
+            { name: "Mt Rugged 4 - Slide Ledge", type: CheckType.NORMAL },
+            { name: "Mt Rugged 4 - Left Ledge Center", type: CheckType.NORMAL },
+            { name: "Mt Rugged 4 - Left Ledge Right", type: CheckType.OVERWORLD_COIN },
+            { name: "Mt Rugged 4 - Left Ledge 3", type: CheckType.OVERWORLD_COIN },
+            { name: "Mt Rugged 4 - Left Ledge 4", type: CheckType.OVERWORLD_COIN },
+            { name: "Mt Rugged 4 - Left Ledge 5", type: CheckType.OVERWORLD_COIN },
+            { name: "Mt Rugged 4 - Left Ledge 6", type: CheckType.OVERWORLD_COIN },
+            { name: "Mt Rugged 4 - Left Ledge 7", type: CheckType.OVERWORLD_COIN },
+            { name: "Mt Rugged 4 - Bottom Left 1", type: CheckType.OVERWORLD_COIN },
+            { name: "Mt Rugged 4 - Bottom Left 2", type: CheckType.OVERWORLD_COIN },
+            { name: "Mt Rugged 4 - Yellow Block Top Left", type: CheckType.COIN_BLOCK },
+            { name: "Mt Rugged 4 - Yellow Block Floating", type: CheckType.NORMAL },
+            { name: "Mt Rugged 4 - Yellow Block Top Right", type: CheckType.NORMAL },
+            { name: "Suspension Bridge - Bottom Of Cliff", type: CheckType.NORMAL },
+        ]
+    },
+    {
+        name: "Dry Dry Outpost", checks: [
+            { name: "Outpost 1 - Composer Lyrics Reward", type: CheckType.NORMAL },
+            { name: "Outpost 1 - Store Legend (RedJar Code)", type: CheckType.NORMAL },
+            { name: "Outpost 1 - Little Mouser Letter Reward", type: CheckType.LETTER_REWARD },
+            { name: "Outpost 1 - Shop Item 1", type: CheckType.SHOP },
+            { name: "Outpost 1 - Shop Item 2", type: CheckType.SHOP },
+            { name: "Outpost 1 - Shop Item 3", type: CheckType.SHOP },
+            { name: "Outpost 1 - Shop Item 4", type: CheckType.SHOP },
+            { name: "Outpost 1 - Shop Item 5", type: CheckType.SHOP },
+            { name: "Outpost 1 - Shop Item 6", type: CheckType.SHOP },
+            { name: "Outpost 1 - In Red Tree", type: CheckType.NORMAL },
+            { name: "Outpost 2 - Merlee Request (Koopa Koot Favor)", type: CheckType.KOOT_FAVOR_ITEM },
+            { name: "Outpost 2 - Moustafa Gift", type: CheckType.NORMAL },
+            { name: "Outpost 2 - Mr. E. Letter Reward", type: CheckType.LETTER_REWARD },
+            { name: "Outpost 2 - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Outpost 2 - Toad House Roof", type: CheckType.NORMAL },
+        ]
+    },
+    {
+        name: "Dry Dry Desert", checks: [
+            { name: "N3W3 - Yellow Block Left", type: CheckType.NORMAL },
+            { name: "N3W3 - Yellow Block Right", type: CheckType.COIN_BLOCK },
+            { name: "N3W1 Ruins Entrance - Radio Trade Event 2 Reward", type: CheckType.TRADE_EVENT },
+            { name: "N3E2 Pokey Army - Behind Cactus", type: CheckType.NORMAL },
+            { name: "N3E3 - In Tree", type: CheckType.FOLIAGE },
+            { name: "N3E3 - In MultiCoinBlock", type: CheckType.MULTICOIN_BLOCK },
+            { name: "N2W3 - Hidden Block", type: CheckType.NORMAL },
+            { name: "N2E1 (Tweester A) - Yellow Block Left", type: CheckType.COIN_BLOCK },
+            { name: "N2E1 (Tweester A) - Yellow Block Right", type: CheckType.NORMAL },
+            { name: "N2E1 (Tweester A) - In MultiCoinBlock", type: CheckType.MULTICOIN_BLOCK },
+            { name: "N1W3 Special Block - Hit Block x1", type: CheckType.NORMAL },
+            { name: "N1W3 Special Block - Hit Block x5", type: CheckType.NORMAL },
+            { name: "N1W3 Special Block - Hit Block x10", type: CheckType.NORMAL },
+            { name: "N1W1 - Yellow Block 1", type: CheckType.COIN_BLOCK },
+            { name: "N1W1 - Yellow Block 2", type: CheckType.COIN_BLOCK },
+            { name: "N1W1 - Yellow Block 3", type: CheckType.COIN_BLOCK },
+            { name: "N1W1 - Yellow Block 4", type: CheckType.COIN_BLOCK },
+            { name: "N1W1 - Yellow Block Center", type: CheckType.NORMAL },
+            { name: "N1E1 Palm Trio - Hidden Block", type: CheckType.NORMAL },
+            { name: "N1E2 - In MultiCoinBlock Center", type: CheckType.MULTICOIN_BLOCK },
+            { name: "N1E2 - In MultiCoinBlock Bottom Right", type: CheckType.MULTICOIN_BLOCK },
+            { name: "N1E3 - In Tree", type: CheckType.FOLIAGE },
+            { name: "W3 Kolorados Camp - In Tree", type: CheckType.FOLIAGE },
+            { name: "Center (Tweester C) - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "E1 Nomadimouse - Nomadimouse Letter Reward", type: CheckType.LETTER_REWARD },
+            { name: "E1 Nomadimouse - In Tree", type: CheckType.FOLIAGE },
+            { name: "E2 - In Tree Far Left", type: CheckType.FOLIAGE },
+            { name: "E3 Outside Outpost - In Tree (Far Left)", type: CheckType.FOLIAGE },
+            { name: "E3 Outside Outpost - In Tree (Second From Left)", type: CheckType.FOLIAGE },
+            { name: "E3 Outside Outpost - In Tree (Fourth From Right)", type: CheckType.FOLIAGE },
+            { name: "E3 Outside Outpost - In Tree (Far Right)", type: CheckType.FOLIAGE },
+            { name: "S1W3 - In MultiCoinBlock Top Left", type: CheckType.MULTICOIN_BLOCK },
+            { name: "S1 - Yellow Block", type: CheckType.COIN_BLOCK },
+            { name: "S1E2 Small Bluffs - On Brick Block", type: CheckType.NORMAL },
+            { name: "S1E2 Small Bluffs - Ontop Of Bluffs", type: CheckType.NORMAL },
+            { name: "S1E3 North of Oasis - Hidden Block", type: CheckType.NORMAL },
+            { name: "S1E3 North of Oasis - Tree Bottom Left", type: CheckType.NORMAL },
+            { name: "S1E3 North of Oasis - Yellow Block", type: CheckType.COIN_BLOCK },
+            { name: "S2W1 - In MultiCoinBlock Top", type: CheckType.MULTICOIN_BLOCK },
+            { name: "S2E2 West of Oasis - Behind Bush", type: CheckType.FOLIAGE },
+            { name: "S2E2 West of Oasis - In Tree", type: CheckType.FOLIAGE },
+            { name: "S2E2 West of Oasis - In MultiCoinBlock", type: CheckType.MULTICOIN_BLOCK },
+            { name: "S2E3 Oasis - In SuperBlock", type: CheckType.SUPER_BLOCK },
+            { name: "S2E3 Oasis - In Fruit Tree (Left)", type: CheckType.NORMAL },
+            { name: "S2E3 Oasis - In Fruit Tree (Right)", type: CheckType.NORMAL },
+            { name: "S2E3 Oasis - In Tree (Far Left)", type: CheckType.FOLIAGE },
+            { name: "S2E3 Oasis - In Tree (Front Right)", type: CheckType.FOLIAGE },
+            { name: "S3W2 Hidden AttackFX - Hidden Block", type: CheckType.NORMAL },
+            { name: "S3E1 - Yellow Block", type: CheckType.COIN_BLOCK },
+            { name: "S3E3 South of Oasis - In Tree (Far Right)", type: CheckType.FOLIAGE },
+            { name: "S3E3 South of Oasis - In MultiCoinBlock Top Left", type: CheckType.MULTICOIN_BLOCK },
+            { name: "S3E3 South of Oasis - In MultiCoinBlock Top Right", type: CheckType.MULTICOIN_BLOCK },
+            { name: "S3E3 South of Oasis - In MultiCoinBlock Right", type: CheckType.MULTICOIN_BLOCK },
+            { name: "S3E3 South of Oasis - In MultiCoinBlock Left", type: CheckType.MULTICOIN_BLOCK },
+            { name: "S3E3 South of Oasis - In MultiCoinBlock Bottom Left", type: CheckType.MULTICOIN_BLOCK },
+            { name: "S3E3 South of Oasis - In MultiCoinBlock Bottom Right", type: CheckType.MULTICOIN_BLOCK },
+        ]
+    },
+    {
+        name: "Dry Dry Ruins", checks: [
+            { name: "Sarcophagus Hall 1 - In Sarcophagus", type: CheckType.NORMAL },
+            { name: "Sand Drainage Room 1 - On The Ground", type: CheckType.NORMAL },
+            { name: "Sand Drainage Room 2 - In The Sand", type: CheckType.NORMAL },
+            { name: "Sand Drainage Room 2 - On Ledge", type: CheckType.NORMAL },
+            { name: "Pyramid Stone Room - On Pedestal", type: CheckType.NORMAL },
+            { name: "Sarcophagus Hall 2 - Pokey Gauntlet Reward", type: CheckType.NORMAL },
+            { name: "Sarcophagus Hall 2 - Behind Hammer Block", type: CheckType.NORMAL },
+            { name: "Super Hammer Room - In Big Chest", type: CheckType.NORMAL },
+            { name: "Super Hammer Room - Hidden Chest", type: CheckType.NORMAL },
+            { name: "Vertical Shaft - In SuperBlock", type: CheckType.SUPER_BLOCK },
+            { name: "Diamond Stone Room - On Pedestal", type: CheckType.NORMAL },
+            { name: "Sand Drainage Room 3 - On Ledge", type: CheckType.NORMAL },
+            { name: "Lunar Stone Room - On Pedestal", type: CheckType.NORMAL },
+        ]
+    },
+    {
+        name: "Forever Forest", checks: [
+            { name: "Tree Face (Bub-ulb) - Bub-ulb Gift", type: CheckType.NORMAL },
+            { name: "Bee Hive (HP Plus) - Central Block", type: CheckType.NORMAL },
+            { name: "Flowers Appear (FP Plus) - Central Block", type: CheckType.NORMAL },
+            { name: "Outside Boos Mansion - In Bush (Back Right)", type: CheckType.FOLIAGE },
+            { name: "Outside Boos Mansion - Yellow Block", type: CheckType.NORMAL },
+            { name: "Exit to Gusty Gulch - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+        ]
+    },
+    {
+        name: "Boos Mansion", checks: [
+            { name: "Foyer - From Franky (Koopa Koot Favor)", type: CheckType.KOOT_FAVOR_ITEM },
+            { name: "Foyer - Franky Letter Reward", type: CheckType.LETTER_REWARD },
+            { name: "Foyer - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Basement Stairs - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Basement - In Crate", type: CheckType.NORMAL },
+            { name: "Basement - Igor Letter Reward", type: CheckType.LETTER_REWARD },
+            { name: "Basement - Shop Item 1", type: CheckType.SHOP },
+            { name: "Basement - Shop Item 2", type: CheckType.SHOP },
+            { name: "Basement - Shop Item 3", type: CheckType.SHOP },
+            { name: "Basement - Shop Item 4", type: CheckType.SHOP },
+            { name: "Basement - Shop Item 5", type: CheckType.SHOP },
+            { name: "Basement - Shop Item 6", type: CheckType.SHOP },
+            { name: "Super Boots Room - In Big Chest", type: CheckType.NORMAL },
+            { name: "Super Boots Room - In Crate", type: CheckType.NORMAL },
+            { name: "Super Boots Room - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Pot Room - In Crate 1", type: CheckType.NORMAL },
+            { name: "Pot Room - In Crate 2", type: CheckType.NORMAL },
+            { name: "Library - In Crate", type: CheckType.NORMAL },
+            { name: "Library - On Bookshelf", type: CheckType.NORMAL },
+            { name: "Record Player Room - In Chest", type: CheckType.NORMAL },
+            { name: "Record Room - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Record Room - Boo Ring Game", type: CheckType.NORMAL },
+            { name: "Lady Bows Room - Bow Partner", type: CheckType.NORMAL },
+        ]
+    },
+    {
+        name: "Gusty Gulch", checks: [
+            { name: "Ghost Town 1 - From Boo (Koopa Koot Favor)", type: CheckType.KOOT_FAVOR_ITEM },
+            { name: "Ghost Town 1 - Yellow Block In House", type: CheckType.COIN_BLOCK },
+            { name: "Wasteland Ascent 1 - On Rock", type: CheckType.NORMAL },
+            { name: "Wasteland Ascent 1 - Infront Of Branch", type: CheckType.NORMAL },
+            { name: "Wasteland Ascent 1 - Yellow Block 1", type: CheckType.COIN_BLOCK },
+            { name: "Wasteland Ascent 1 - Yellow Block 2", type: CheckType.COIN_BLOCK },
+            { name: "Wasteland Ascent 1 - Yellow Block Right", type: CheckType.NORMAL },
+            { name: "Wasteland Ascent 2 - Behind Rock", type: CheckType.NORMAL },
+            { name: "Wasteland Ascent 2 - In MultiCoinBlock", type: CheckType.MULTICOIN_BLOCK },
+            { name: "Wasteland Ascent 2 - Yellow Block Left", type: CheckType.NORMAL },
+            { name: "Wasteland Ascent 2 - Yellow Block Right", type: CheckType.COIN_BLOCK },
+        ]
+    },
+    {
+        name: "Tubbas Castle", checks: [
+            { name: "Covered Tables Room (1F) - On Table", type: CheckType.NORMAL },
+            { name: "Study (1F) - On Table", type: CheckType.NORMAL },
+            { name: "Table/Clock Room (1/2F) - On Table", type: CheckType.NORMAL },
+            { name: "Basement - In Chest", type: CheckType.NORMAL },
+            { name: "Stairs to Basement - In SuperBlock", type: CheckType.SUPER_BLOCK },
+            { name: "Spike Trap Room (2F) - In Chest", type: CheckType.NORMAL },
+            { name: "Hidden Bedroom (2F) - In Hidden Room", type: CheckType.NORMAL },
+            { name: "Hidden Bedroom (2F) - On Bed 1", type: CheckType.OVERWORLD_COIN },
+            { name: "Hidden Bedroom (2F) - On Bed 2", type: CheckType.OVERWORLD_COIN },
+            { name: "Hidden Bedroom (2F) - On Bed 3", type: CheckType.OVERWORLD_COIN },
+            { name: "Hidden Bedroom (2F) - On Bed 4", type: CheckType.OVERWORLD_COIN },
+            { name: "Hidden Bedroom (2F) - On Bed 5", type: CheckType.OVERWORLD_COIN },
+            { name: "Hidden Bedroom (2F) - On Bed 6", type: CheckType.OVERWORLD_COIN },
+            { name: "Stairs to Third Floor - Yellow Block", type: CheckType.NORMAL },
+            { name: "Sleeping Clubbas Room (3F) - On Pedestal", type: CheckType.NORMAL },
+        ]
+    },
+    {
+        name: "Shy Guys Toybox", checks: [
+            { name: "BLU Station - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "BLU Station - Hidden Block", type: CheckType.NORMAL },
+            { name: "BLU Anti-Guy Hall - In Chest", type: CheckType.NORMAL },
+            { name: "BLU Anti-Guy Hall - Hidden Block", type: CheckType.NORMAL },
+            { name: "BLU Anti-Guy Hall - Yellow Block", type: CheckType.COIN_BLOCK },
+            { name: "BLU Large Playroom - Hidden Block 1", type: CheckType.NORMAL },
+            { name: "BLU Large Playroom - Hidden Block 2", type: CheckType.NORMAL },
+            { name: "BLU Large Playroom - Calculator Thief 1", type: CheckType.NORMAL },
+            { name: "BLU Large Playroom - Calculator Thief 2", type: CheckType.NORMAL },
+            { name: "BLU Large Playroom - Shy Guy 2", type: CheckType.NORMAL },
+            { name: "BLU Large Playroom - Shy Guy 3", type: CheckType.NORMAL },
+            { name: "BLU Large Playroom - Shy Guy 4", type: CheckType.NORMAL },
+            { name: "BLU Large Playroom - Shy Guy 5", type: CheckType.NORMAL },
+            { name: "BLU Block City - In Chest", type: CheckType.NORMAL },
+            { name: "BLU Block City - Infront Of Chest", type: CheckType.NORMAL },
+            { name: "BLU Block City - Midair 1", type: CheckType.OVERWORLD_COIN },
+            { name: "BLU Block City - Midair 2", type: CheckType.OVERWORLD_COIN },
+            { name: "BLU Block City - Midair 3", type: CheckType.OVERWORLD_COIN },
+            { name: "BLU Block City - Midair 4", type: CheckType.OVERWORLD_COIN },
+            { name: "BLU Block City - Midair 5", type: CheckType.OVERWORLD_COIN },
+            { name: "BLU Block City - Midair 6", type: CheckType.OVERWORLD_COIN },
+            { name: "BLU Block City - Midair 7", type: CheckType.OVERWORLD_COIN },
+            { name: "BLU Block City - Midair 8", type: CheckType.OVERWORLD_COIN },
+            { name: "BLU Block City - On Building", type: CheckType.NORMAL },
+            { name: "BLU Block City - Behind Building Block", type: CheckType.NORMAL },
+            { name: "BLU Block City - Yellow Block 1", type: CheckType.COIN_BLOCK },
+            { name: "BLU Block City - Yellow Block 2", type: CheckType.COIN_BLOCK },
+            { name: "BLU Block City - Yellow Block On Ledge", type: CheckType.NORMAL },
+            { name: "PNK Station - In Chest", type: CheckType.NORMAL },
+            { name: "PNK Station - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "PNK Station - Hidden Block", type: CheckType.NORMAL },
+            { name: "PNK Tracks Hallway - Yellow Block South", type: CheckType.COIN_BLOCK },
+            { name: "PNK Tracks Hallway - Yellow Block North 1", type: CheckType.COIN_BLOCK },
+            { name: "PNK Tracks Hallway - Yellow Block North 2", type: CheckType.COIN_BLOCK },
+            { name: "PNK Tracks Hallway - In MultiCoinBlock", type: CheckType.MULTICOIN_BLOCK },
+            { name: "PNK Gourmet Guy Crossing - Hidden Block Right", type: CheckType.NORMAL },
+            { name: "PNK Gourmet Guy Crossing - Hidden Block Left", type: CheckType.NORMAL },
+            { name: "PNK Gourmet Guy Crossing - Gourmet Guy Reward", type: CheckType.NORMAL },
+            { name: "PNK Gourmet Guy Crossing - Yellow Block 1", type: CheckType.COIN_BLOCK },
+            { name: "PNK Gourmet Guy Crossing - Yellow Block 2", type: CheckType.COIN_BLOCK },
+            { name: "PNK Playhouse - In Chest (Far Right)", type: CheckType.NORMAL },
+            { name: "PNK Playhouse - In Chest (Top Left)", type: CheckType.NORMAL },
+            { name: "PNK Playhouse - In Chest (Right)", type: CheckType.NORMAL },
+            { name: "PNK Playhouse - Infront Of Chest (Right)", type: CheckType.NORMAL },
+            { name: "PNK Playhouse - Yellow Block", type: CheckType.COIN_BLOCK },
+            { name: "GRN Station - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "GRN Station - Hidden Block", type: CheckType.NORMAL },
+            { name: "GRN Treadmills/Slot Machine - In Chest", type: CheckType.NORMAL },
+            { name: "GRN Treadmills/Slot Machine - Infront Of Chest", type: CheckType.NORMAL },
+            { name: "GRN Treadmills/Slot Machine - On Treadmill 1", type: CheckType.OVERWORLD_COIN },
+            { name: "GRN Treadmills/Slot Machine - On Treadmill 2", type: CheckType.OVERWORLD_COIN },
+            { name: "GRN Treadmills/Slot Machine - On Treadmill 3", type: CheckType.OVERWORLD_COIN },
+            { name: "GRN Treadmills/Slot Machine - On Treadmill 4", type: CheckType.OVERWORLD_COIN },
+            { name: "GRN Treadmills/Slot Machine - On Treadmill 5", type: CheckType.OVERWORLD_COIN },
+            { name: "GRN Treadmills/Slot Machine - On Treadmill 6", type: CheckType.OVERWORLD_COIN },
+            { name: "GRN Treadmills/Slot Machine - Hidden Room Center", type: CheckType.NORMAL },
+            { name: "GRN Treadmills/Slot Machine - Hidden Room 1", type: CheckType.OVERWORLD_COIN },
+            { name: "GRN Treadmills/Slot Machine - Hidden Room 2", type: CheckType.OVERWORLD_COIN },
+            { name: "GRN Treadmills/Slot Machine - Hidden Room 3", type: CheckType.OVERWORLD_COIN },
+            { name: "GRN Treadmills/Slot Machine - Hidden Room 4", type: CheckType.OVERWORLD_COIN },
+            { name: "GRN Treadmills/Slot Machine - Hidden Room 5", type: CheckType.OVERWORLD_COIN },
+            { name: "GRN Treadmills/Slot Machine - Hidden Room 6", type: CheckType.OVERWORLD_COIN },
+            { name: "GRN Treadmills/Slot Machine - Defeat Shy Guy", type: CheckType.NORMAL },
+            { name: "GRN Treadmills/Slot Machine - In MultiCoinBlock", type: CheckType.MULTICOIN_BLOCK },
+            { name: "RED Station - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "RED Station - Hidden Block", type: CheckType.NORMAL },
+            { name: "RED Moving Platforms - Hidden Block Center", type: CheckType.NORMAL },
+            { name: "RED Moving Platforms - Hidden Block Right", type: CheckType.NORMAL },
+            { name: "RED Moving Platforms - Hidden Block Left", type: CheckType.NORMAL },
+            { name: "RED Moving Platforms - In SuperBlock", type: CheckType.SUPER_BLOCK },
+            { name: "RED Moving Platforms - In MultiCoinBlock", type: CheckType.MULTICOIN_BLOCK },
+            { name: "RED Moving Platforms - Yellow Block 1", type: CheckType.COIN_BLOCK },
+            { name: "RED Moving Platforms - Yellow Block 2", type: CheckType.COIN_BLOCK },
+            { name: "RED Lantern Ghost - Watt Partner", type: CheckType.NORMAL },
+            { name: "RED Boss Barricade - Hidden Block Left", type: CheckType.COIN_BLOCK },
+            { name: "RED Boss Barricade - On Brick Block", type: CheckType.NORMAL },
+            { name: "RED Boss Barricade - Yellow Block Right", type: CheckType.NORMAL },
+        ]
+    },
+    {
+        name: "Jade Jungle", checks: [
+            { name: "Whale Cove - Over Flower 1", type: CheckType.OVERWORLD_COIN },
+            { name: "Whale Cove - Over Flower 2", type: CheckType.OVERWORLD_COIN },
+            { name: "Whale Cove - Behind Bush", type: CheckType.FOLIAGE },
+            { name: "Whale Cove - In Palm Tree", type: CheckType.FOLIAGE },
+            { name: "Beach - Hidden Block Left", type: CheckType.NORMAL },
+            { name: "Beach - Hidden Block Right", type: CheckType.NORMAL },
+            { name: "Beach - On The Rocks", type: CheckType.NORMAL },
+            { name: "Beach - Over Flower 1", type: CheckType.OVERWORLD_COIN },
+            { name: "Beach - Over Flower 2", type: CheckType.OVERWORLD_COIN },
+            { name: "Beach - In Palm Tree 1", type: CheckType.FOLIAGE },
+            { name: "Beach - In Palm Tree 2", type: CheckType.FOLIAGE },
+            { name: "Beach - In Palm Tree 3", type: CheckType.FOLIAGE },
+            { name: "Beach - In Palm Tree 4", type: CheckType.FOLIAGE },
+            { name: "Beach - In Palm Tree 5", type: CheckType.FOLIAGE },
+            { name: "Beach - In Palm Tree 6", type: CheckType.FOLIAGE },
+            { name: "Beach - In Palm Tree 6 2", type: CheckType.FOLIAGE },
+            { name: "Village Cove - Village Leader Reward", type: CheckType.NORMAL },
+            { name: "Village Cove - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Village Cove - In Palm Tree Left", type: CheckType.FOLIAGE },
+            { name: "Village Cove - In Palm Tree Right", type: CheckType.FOLIAGE },
+            { name: "Village Buildings - Kolorado Volcano Vase Reward", type: CheckType.NORMAL },
+            { name: "Village Buildings - Yellow Yoshi Food Reward", type: CheckType.NORMAL },
+            { name: "Village Buildings - Red Yoshi Kid Letter Reward", type: CheckType.LETTER_REWARD },
+            { name: "Village Buildings - Shop Item 1", type: CheckType.SHOP },
+            { name: "Village Buildings - Shop Item 2", type: CheckType.SHOP },
+            { name: "Village Buildings - Shop Item 3", type: CheckType.SHOP },
+            { name: "Village Buildings - Shop Item 4", type: CheckType.SHOP },
+            { name: "Village Buildings - Shop Item 5", type: CheckType.SHOP },
+            { name: "Village Buildings - Shop Item 6", type: CheckType.SHOP },
+            { name: "Village Buildings - In Palm Tree", type: CheckType.FOLIAGE },
+            { name: "Path to the Volcano - Raphael Gift", type: CheckType.NORMAL },
+            { name: "Path to the Volcano - Behind Tree", type: CheckType.NORMAL },
+            { name: "SE Jungle (Quake Hammer) - Bush (Bottom Right)", type: CheckType.FOLIAGE },
+            { name: "SE Jungle (Quake Hammer) - Bush (Bottom Left)", type: CheckType.FOLIAGE },
+            { name: "SE Jungle (Quake Hammer) - Red Block", type: CheckType.NORMAL },
+            { name: "SE Jungle (Quake Hammer) - In Tree (Right)", type: CheckType.FOLIAGE },
+            { name: "Sushi Tree - In Volcano Chest", type: CheckType.NORMAL },
+            { name: "Sushi Tree - On Island", type: CheckType.NORMAL },
+            { name: "Sushi Tree - Sushie Partner", type: CheckType.NORMAL },
+            { name: "Sushi Tree - In Island Tree", type: CheckType.FOLIAGE },
+            { name: "SW Jungle (Super Block) - Bush (Top Right)", type: CheckType.FOLIAGE },
+            { name: "SW Jungle (Super Block) - Bush (Bottom Left)", type: CheckType.FOLIAGE },
+            { name: "SW Jungle (Super Block) - Hidden Block", type: CheckType.NORMAL },
+            { name: "SW Jungle (Super Block) - Underwater 1", type: CheckType.OVERWORLD_COIN },
+            { name: "SW Jungle (Super Block) - Underwater 2", type: CheckType.OVERWORLD_COIN },
+            { name: "SW Jungle (Super Block) - Underwater 3", type: CheckType.OVERWORLD_COIN },
+            { name: "SW Jungle (Super Block) - In SuperBlock", type: CheckType.SUPER_BLOCK },
+            { name: "SW Jungle (Super Block) - In Tree (Top)", type: CheckType.FOLIAGE },
+            { name: "SW Jungle (Super Block) - In Tree (Right)", type: CheckType.FOLIAGE },
+            { name: "NW Jungle (Large Ledge) - Bush 1", type: CheckType.FOLIAGE },
+            { name: "NW Jungle (Large Ledge) - Bush 2", type: CheckType.FOLIAGE },
+            { name: "NW Jungle (Large Ledge) - In Tree On Ledge", type: CheckType.FOLIAGE },
+            { name: "NW Jungle (Large Ledge) - In Tree Right", type: CheckType.FOLIAGE },
+            { name: "Western Dead End - Underwater", type: CheckType.OVERWORLD_COIN },
+            { name: "NE Jungle (Raven Statue) - Underwater", type: CheckType.OVERWORLD_COIN },
+            { name: "NE Jungle (Raven Statue) - In Tree (Top Left)", type: CheckType.FOLIAGE },
+            { name: "Small Jungle Ledge - In Tree", type: CheckType.FOLIAGE },
+            { name: "Deep Jungle 1 - Hidden Block", type: CheckType.NORMAL },
+            { name: "Deep Jungle 1 - In Tree (Vine)", type: CheckType.NORMAL },
+            { name: "Deep Jungle 1 - In Tree (Hit)", type: CheckType.FOLIAGE },
+            { name: "Deep Jungle 2 (Block Puzzle) - Hidden Block", type: CheckType.NORMAL },
+            { name: "Deep Jungle 2 (Block Puzzle) - In Tree (Left)", type: CheckType.FOLIAGE },
+            { name: "Deep Jungle 3 - Tree Vine Second Left", type: CheckType.NORMAL },
+            { name: "Deep Jungle 3 - Tree Vine Far Right", type: CheckType.NORMAL },
+            { name: "Deep Jungle 4 (Ambush) - Hidden Panel", type: CheckType.FOLIAGE },
+            { name: "Deep Jungle 4 (Ambush) - In Tree (Right)", type: CheckType.FOLIAGE },
+            { name: "Great Tree Vine Ascent - End Of Vine", type: CheckType.NORMAL },
+        ]
+    },
+    {
+        name: "Mt Lavalava", checks: [
+            { name: "Central Cavern - On Stone Pillar", type: CheckType.NORMAL },
+            { name: "Central Cavern - On Brick Block", type: CheckType.NORMAL },
+            { name: "Central Cavern - Yellow Block 1", type: CheckType.COIN_BLOCK },
+            { name: "Central Cavern - Yellow Block 2", type: CheckType.COIN_BLOCK },
+            { name: "Central Cavern - Yellow Block 3", type: CheckType.COIN_BLOCK },
+            { name: "Central Cavern - Yellow Block 4", type: CheckType.COIN_BLOCK },
+            { name: "Fire Bar Bridge - In SuperBlock", type: CheckType.SUPER_BLOCK },
+            { name: "Flowing Lava Puzzle - Hidden Block", type: CheckType.NORMAL },
+            { name: "Ultra Hammer Room - In Big Chest", type: CheckType.NORMAL },
+            { name: "Dizzy Stomp Room - In Chest", type: CheckType.NORMAL },
+            { name: "Zipline Cavern - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Zipline Cavern - In SuperBlock", type: CheckType.SUPER_BLOCK },
+            { name: "Boss Antechamber - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Boss Room - Yellow Block Left", type: CheckType.NORMAL },
+            { name: "Boss Room - Yellow Block Right", type: CheckType.NORMAL },
+        ]
+    },
+    {
+        name: "Flower Fields", checks: [
+            { name: "(NE) Elevators - Stomp On Ledge", type: CheckType.NORMAL },
+            { name: "(NE) Elevators - Leftside Vine", type: CheckType.NORMAL },
+            { name: "(NE) Elevators - In SuperBlock", type: CheckType.SUPER_BLOCK },
+            { name: "(NE) Fallen Logs - Hidden Block", type: CheckType.NORMAL },
+            { name: "(NE) Fallen Logs - In The Flowers", type: CheckType.NORMAL },
+            { name: "(East) Triple Tree Path - Leftmost Vine", type: CheckType.NORMAL },
+            { name: "(East) Triple Tree Path - Tree Puzzle Reward", type: CheckType.NORMAL },
+            { name: "(East) Petunias Field - Petunia Gift", type: CheckType.NORMAL },
+            { name: "(East) Petunias Field - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "(East) Petunias Field - In Tree 1", type: CheckType.FOLIAGE },
+            { name: "(East) Petunias Field - In Tree 2", type: CheckType.FOLIAGE },
+            { name: "(East) Old Well - Well BlueBerry Reward", type: CheckType.NORMAL },
+            { name: "(SE) Briar Platforming - In The Flowers", type: CheckType.NORMAL },
+            { name: "(SE) Briar Platforming - Left Side Vine", type: CheckType.NORMAL },
+            { name: "(SE) Briar Platforming - In SuperBlock", type: CheckType.SUPER_BLOCK },
+            { name: "(SE) Briar Platforming - In Tree 1", type: CheckType.FOLIAGE },
+            { name: "(SE) Briar Platforming - In Tree 2", type: CheckType.FOLIAGE },
+            { name: "(SE) Water Level Room - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "(SE) Water Level Room - Hidden Block", type: CheckType.NORMAL },
+            { name: "(SE) Water Level Room - In Tree 1", type: CheckType.FOLIAGE },
+            { name: "(SE) Water Level Room - In Tree 2", type: CheckType.FOLIAGE },
+            { name: "(SE) Water Level Room - Yellow Block", type: CheckType.NORMAL },
+            { name: "(SE) Lilys Fountain - Lily Reward For WaterStone", type: CheckType.NORMAL },
+            { name: "(SE) Lilys Fountain - In Tree", type: CheckType.FOLIAGE },
+            { name: "(SW) Path to Crystal Tree - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "(SW) Path to Crystal Tree - Central Vine", type: CheckType.NORMAL },
+            { name: "(SW) Path to Crystal Tree - In Tree 1", type: CheckType.FOLIAGE },
+            { name: "(SW) Path to Crystal Tree - In Tree 2", type: CheckType.FOLIAGE },
+            { name: "(SW) Posie and Crystal Tree - Posie Gift 1", type: CheckType.NORMAL },
+            { name: "(SW) Posie and Crystal Tree - Posie Gift 2", type: CheckType.NORMAL },
+            { name: "(West) Path to Maze - Upper Hidden Block", type: CheckType.NORMAL },
+            { name: "(West) Path to Maze - Lower Hidden Block", type: CheckType.COIN_BLOCK },
+            { name: "(West) Maze - In MultiCoinBlock", type: CheckType.MULTICOIN_BLOCK },
+            { name: "(West) Rosies Trellis - Rosie Gift", type: CheckType.NORMAL },
+            { name: "(NW) Bubble Flower - On Ledge", type: CheckType.NORMAL },
+            { name: "(NW) Bubble Flower - Right Vine", type: CheckType.NORMAL },
+            { name: "(NW) Lakilester - Cage Under Rock", type: CheckType.NORMAL },
+            { name: "(NW) Lakilester - In The Flowers", type: CheckType.NORMAL },
+            { name: "(NW) Lakilester - Lakilester Partner", type: CheckType.NORMAL },
+            { name: "Cloudy Climb - On Cloud", type: CheckType.NORMAL },
+        ]
+    },
+    {
+        name: "Shiver Region", checks: [
+            { name: "Shiver City Center - Sleep At Toad House", type: CheckType.NORMAL },
+            { name: "Shiver City Center - Snowmen Gift 1", type: CheckType.NORMAL },
+            { name: "Shiver City Center - Snowmen Gift 2", type: CheckType.NORMAL },
+            { name: "Shiver City Center - Snowmen Gift 3", type: CheckType.NORMAL },
+            { name: "Shiver City Center - Snowmen Gift 4", type: CheckType.NORMAL },
+            { name: "Shiver City Center - Snowmen Gift 5", type: CheckType.NORMAL },
+            { name: "Shiver City Center - Shop Item 1", type: CheckType.SHOP },
+            { name: "Shiver City Center - Shop Item 2", type: CheckType.SHOP },
+            { name: "Shiver City Center - Shop Item 3", type: CheckType.SHOP },
+            { name: "Shiver City Center - Shop Item 4", type: CheckType.SHOP },
+            { name: "Shiver City Center - Shop Item 5", type: CheckType.SHOP },
+            { name: "Shiver City Center - Shop Item 6", type: CheckType.SHOP },
+            { name: "Shiver City Mayor Area - Chest In House", type: CheckType.NORMAL },
+            { name: "Shiver City Mayor Area - Mayor Penguin Gift", type: CheckType.NORMAL },
+            { name: "Shiver City Mayor Area - Mayor Penguin Letter Reward", type: CheckType.LETTER_REWARD },
+            { name: "Shiver City Mayor Area - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Shiver City Pond Area - In Frozen Pond", type: CheckType.NORMAL },
+            { name: "Shiver Snowfield - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Shiver Snowfield - Behind Tree Right", type: CheckType.NORMAL },
+            { name: "Shiver Snowfield - In Tree Left", type: CheckType.NORMAL },
+            { name: "Path to Starborn Valley - Hidden Block", type: CheckType.NORMAL },
+            { name: "Path to Starborn Valley - Behind Icicle", type: CheckType.NORMAL },
+            { name: "Starborn Valley - Merle Gift", type: CheckType.NORMAL },
+            { name: "Starborn Valley - Frost T. Letter Reward", type: CheckType.LETTER_REWARD },
+            { name: "Shiver Mountain Passage - Hidden Block", type: CheckType.NORMAL },
+            { name: "Shiver Mountain Hills - Bottom Path", type: CheckType.NORMAL },
+            { name: "Shiver Mountain Hills - In SuperBlock", type: CheckType.SUPER_BLOCK },
+            { name: "Shiver Mountain Tunnel - Socket 1", type: CheckType.NORMAL },
+            { name: "Shiver Mountain Tunnel - Socket 2", type: CheckType.NORMAL },
+            { name: "Shiver Mountain Tunnel - Socket 3", type: CheckType.NORMAL },
+            { name: "Shiver Mountain Peaks - Left Ledge", type: CheckType.NORMAL },
+            { name: "Shiver Mountain Peaks - Red Block", type: CheckType.NORMAL },
+            { name: "Merlars Sanctuary - On Pedestal", type: CheckType.NORMAL },
+        ]
+    },
+    {
+        name: "Crystal Palace", checks: [
+            { name: "Blue Key Room - In Chest", type: CheckType.NORMAL },
+            { name: "Red Key Room - In Chest", type: CheckType.NORMAL },
+            { name: "Reflected Save Room - Yellow Block", type: CheckType.NORMAL },
+            { name: "Shooting Star Room - On The Ground", type: CheckType.NORMAL },
+            { name: "P-Down, D-Up Room - In Chest", type: CheckType.NORMAL },
+            { name: "Blue Mirror Hall 2 - In MultiCoinBlock Front", type: CheckType.MULTICOIN_BLOCK },
+            { name: "Blue Mirror Hall 2 - In MultiCoinBlock Back", type: CheckType.MULTICOIN_BLOCK },
+            { name: "Star Piece Cave - On The Ground", type: CheckType.NORMAL },
+            { name: "Triple Dip Room - In Chest", type: CheckType.NORMAL },
+            { name: "Huge Statue Room - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Huge Statue Room - Yellow Block", type: CheckType.NORMAL },
+            { name: "Palace Key Room - In Chest", type: CheckType.NORMAL },
+            { name: "Small Statue Room - Hidden Panel", type: CheckType.HIDDEN_PANEL },
+            { name: "Small Statue Room - Hidden Block", type: CheckType.NORMAL },
+            { name: "P-Up, D-Down Room - In Chest", type: CheckType.NORMAL },
+        ]
+    },
+    {
+        name: "Bowsers Castle", checks: [
+            { name: "Front Door Exterior - Red Block", type: CheckType.NORMAL },
+            { name: "Lower Jail - In Crate 1", type: CheckType.NORMAL },
+            { name: "Lower Jail - In Crate 2", type: CheckType.NORMAL },
+            { name: "Outside Lower Jail - Defeat Koopatrol Reward", type: CheckType.NORMAL },
+            { name: "Outside Lower Jail - Yellow Block", type: CheckType.NORMAL },
+            { name: "Lava Key Room - In Chest", type: CheckType.NORMAL },
+            { name: "Lava Channel 3 - On Island 1", type: CheckType.NORMAL },
+            { name: "Lava Channel 3 - On Island 2", type: CheckType.NORMAL },
+            { name: "Dark Cave 1 - Yellow Block", type: CheckType.NORMAL },
+            { name: "Dark Cave 2 - Yellow Block", type: CheckType.NORMAL },
+            { name: "East Upper Jail - Defeat Koopatrol Reward", type: CheckType.NORMAL },
+            { name: "Item Shop - Shop Item 1", type: CheckType.SHOP },
+            { name: "Item Shop - Shop Item 2", type: CheckType.SHOP },
+            { name: "Item Shop - Shop Item 3", type: CheckType.SHOP },
+            { name: "Item Shop - Shop Item 4", type: CheckType.SHOP },
+            { name: "Item Shop - Shop Item 5", type: CheckType.SHOP },
+            { name: "Item Shop - Shop Item 6", type: CheckType.SHOP },
+            { name: "Left Water Puzzle - Top Left Ledge", type: CheckType.NORMAL },
+            { name: "Right Water Puzzle - Hidden Block", type: CheckType.NORMAL },
+            { name: "Room with Hidden Door 1 - Hidden Block", type: CheckType.NORMAL },
+            { name: "Room with Hidden Door 1 - Yellow Block", type: CheckType.NORMAL },
+            { name: "Hidden Key Room - On The Ground", type: CheckType.NORMAL },
+            { name: "Battlement - On Ledge", type: CheckType.NORMAL },
+            { name: "Battlement - Yellow Block Left", type: CheckType.NORMAL },
+            { name: "Battlement - Yellow Block Center", type: CheckType.NORMAL },
+            { name: "Battlement - Yellow Block Right", type: CheckType.NORMAL },
+            { name: "West Upper Jail - Defeat Koopatrol Reward", type: CheckType.NORMAL },
+            { name: "Ultra Shroom Room - On The Ground", type: CheckType.NORMAL },
+            { name: "Castle Key Room - On The Ground", type: CheckType.NORMAL },
+        ]
+    },
+    {
+        name: "Peachs Castle", checks: [
+            { name: "Hijacked Castle Entrance - Hidden Block", type: CheckType.NORMAL },
+            { name: "Guest Room (1F) - In Chest", type: CheckType.NORMAL },
+            { name: "Library (2F) - Upper Level", type: CheckType.NORMAL },
+            { name: "Library (2F) - Between Bookshelves", type: CheckType.NORMAL },
+            { name: "Storeroom (2F) - On The Ground", type: CheckType.NORMAL },
+        ]
+    }];
+
+export const PLANDO_ITEMS_LIST: Array<string> = [
+    "TRAP",
+    "NonProgression",
+    "Consumable",
+    "AllorNothing",
+    "Apple",
+    "ApplePie",
+    "Artifact",
+    "AttackFXA",
+    "AttackFXB",
+    "AttackFXC",
+    "AttackFXD",
+    "AttackFXE",
+    "AttackFXF",
+    "AutoJump",
+    "AutoMultibounce",
+    "AutoSmash",
+    "Berserker",
+    "BigCookie",
+    "BlandMeal",
+    "BlueBerry",
+    "BlueKey",
+    "BoiledEgg",
+    "Bombette",
+    "BombetteUpgrade",
+    "BooPortrait",
+    "BooRecord",
+    "BooWeight",
+    "Bow",
+    "BowUpgrade",
+    "BowserCastleKey",
+    "BubbleBerry",
+    "BumpAttack",
+    "Cake",
+    "CakeMix",
+    "Calculator",
+    "ChillOut",
+    "CloseCall",
+    "CocoPop",
+    "Coconut",
+    "Coin",
+    "Cookbook",
+    "CrazyHeart",
+    "CrystalBall",
+    "CrystalBerry",
+    "CrystalPalaceKey",
+    "DDownJump",
+    "DDownPound",
+    "DamageDodge",
+    "DeepFocus",
+    "DefendPlus",
+    "DeluxeFeast",
+    "DiamondStone",
+    "Dictionary",
+    "Diploma",
+    "DizzyAttack",
+    "DizzyDial",
+    "DizzyStomp",
+    "DodgeMaster",
+    "Dolly",
+    "DoubleDip",
+    "DriedFruit",
+    "DriedPasta",
+    "DriedShroom",
+    "DustyHammer",
+    "Egg",
+    "EggMissile",
+    "ElectroPop",
+    "FPPlus",
+    "FeelingFine",
+    "FertileSoil",
+    "FireFlower",
+    "FirePop",
+    "FireShield",
+    "FirstAttack",
+    "FirstDegreeCard",
+    "FlowerFanatic",
+    "FlowerFinder",
+    "FlowerSaver",
+    "ForestPass",
+    "FourthDegreeCard",
+    "FriedEgg",
+    "FriedShroom",
+    "FrightJar",
+    "FrozenFries",
+    "FryingPan",
+    "GoldCredit",
+    "Goombario",
+    "GoombarioUpgrade",
+    "Goomnut",
+    "GroupFocus",
+    "HPDrain",
+    "HPPlus",
+    "HammerThrow",
+    "HappyFlower",
+    "HappyHeart",
+    "HealthyHealthy",
+    "HealthyJuice",
+    "HeartFinder",
+    "HoneyCandy",
+    "HoneyShroom",
+    "HoneySuper",
+    "HoneySyrup",
+    "HoneyUltra",
+    "HotShroom",
+    "HustleDrink",
+    "ISpy",
+    "IcePower",
+    "IcedPotato",
+    "InsecticideHerb",
+    "ItemPouch",
+    "JadeRaven",
+    "JamminJelly",
+    "JellyPop",
+    "JellyShroom",
+    "JellySuper",
+    "JellyUltra",
+    "JumpCharge",
+    "KookyCookie",
+    "KoopaFortressKey",
+    "KoopaLeaf",
+    "KoopaTea",
+    "Koopasta",
+    "Kooper",
+    "KooperShell",
+    "KooperUpgrade",
+    "KootEmptyWallet",
+    "KootGlasses",
+    "KootKoopaLegends",
+    "KootLuigiAutograph",
+    "KootMerluvleeAutograph",
+    "KootOldPhoto",
+    "KootPackage",
+    "KootRedJar",
+    "KootTheTape",
+    "Lakilester",
+    "LakilesterUpgrade",
+    "LastStand",
+    "Lemon",
+    "LemonCandy",
+    "Letter (To Dane T.)",
+    "Letter (To Fice T.)",
+    "Letter (To Fishmael)",
+    "Letter (To Franky)",
+    "Letter (To Frost T.)",
+    "Letter (To Goompa)",
+    "Letter (To Goompapa) 1",
+    "Letter (To Goompapa) 2",
+    "Letter (To Igor)",
+    "Letter (To Kolorado)",
+    "Letter (To Koover) 1",
+    "Letter (To Koover) 2",
+    "Letter (To Little Mouser)",
+    "Letter (To Mayor Penguin)",
+    "Letter (To Merlon)",
+    "Letter (To Merlow)",
+    "Letter (To Minh T.)",
+    "Letter (To Miss T.)",
+    "Letter (To Mort T.)",
+    "Letter (To Mr. E.)",
+    "Letter (To Muss T.)",
+    "Letter (To Nomadimouse)",
+    "Letter (To Red Yoshi Kid)",
+    "Letter (To Russ T.)",
+    "LifeShroom",
+    "Lime",
+    "LimeCandy",
+    "LuckyDay",
+    "LunarStone",
+    "Lyrics",
+    "MagicalBean",
+    "MagicalSeed1",
+    "MagicalSeed2",
+    "MagicalSeed3",
+    "MagicalSeed4",
+    "Mailbag",
+    "MapleShroom",
+    "MapleSuper",
+    "MapleSyrup",
+    "MapleUltra",
+    "MegaHPDrain",
+    "MegaJump",
+    "MegaQuake",
+    "MegaRush",
+    "MegaSmash",
+    "Melody",
+    "Melon",
+    "MiniJumpCharge",
+    "MiniSmashCharge",
+    "MiracleWater",
+    "Mistake",
+    "MoneyMoney",
+    "Multibounce",
+    "Mushroom",
+    "Mystery",
+    "MysteryNote",
+    "NuttyCake",
+    "OddKey",
+    "PDownDUp",
+    "POWBlock",
+    "PUpDDown",
+    "Parakarry",
+    "ParakarryUpgrade",
+    "PayOff",
+    "Pebble",
+    "Peekaboo",
+    "PotatoSalad",
+    "PowerBounce",
+    "PowerJump",
+    "PowerPlus",
+    "PowerQuake",
+    "PowerRush",
+    "PowerSmash",
+    "PowerStar",
+    "PrettyLucky",
+    "PrisonKey",
+    "ProgressiveBoots",
+    "ProgressiveHammer",
+    "ProgressiveJumpCharge",
+    "ProgressivePowerJump",
+    "ProgressivePowerSmash",
+    "ProgressiveQuakeHammer",
+    "ProgressiveSmashCharge",
+    "PulseStone",
+    "PyramidStone",
+    "QuakeBounce",
+    "QuakeHammer",
+    "QuickChange",
+    "RedBerry",
+    "RedKey",
+    "Refund",
+    "RepelGel",
+    "RightOn",
+    "RuinsKey",
+    "RunawayPay",
+    "SJumpCharge",
+    "SSmashChg",
+    "SecondDegreeCard",
+    "ShootingStar",
+    "ShrinkSmash",
+    "ShrinkStomp",
+    "ShroomCake",
+    "ShroomSteak",
+    "SilverCredit",
+    "SleepStomp",
+    "SleepySheep",
+    "SlowGo",
+    "SmashCharge",
+    "SnowmanBucket",
+    "SnowmanDoll",
+    "SnowmanScarf",
+    "Spaghetti",
+    "SpecialShake",
+    "SpeedySpin",
+    "SpicySoup",
+    "SpikeShield",
+    "SpinAttack",
+    "SpinSmash",
+    "StarBeam",
+    "StarPiece",
+    "StarStone",
+    "StinkyHerb",
+    "StoneCap",
+    "StopWatch",
+    "StoreroomKey",
+    "StrangeCake",
+    "StrangeLeaf",
+    "SuperFocus",
+    "SuperJump",
+    "SuperShroom",
+    "SuperSmash",
+    "SuperSoda",
+    "Sushie",
+    "SushieUpgrade",
+    "SweetShroom",
+    "TastyTonic",
+    "ThirdDegreeCard",
+    "ThreeStarPieces",
+    "ThunderBolt",
+    "ThunderRage",
+    "ToyTrain",
+    "TripleDip",
+    "TubbaCastleKey",
+    "UltraShroom",
+    "UltraStone",
+    "VolcanoVase",
+    "VoltShroom",
+    "WarehouseKey",
+    "WaterStone",
+    "Watt",
+    "WattUpgrade",
+    "WhackasBump",
+    "YellowBerry",
+    "YoshiCookie",
+    "YummyMeal",
+    "ZapTap"
+];
+
+export const KEY_TO_DUNGEON = {
+    "BlueKey": "Crystal Palace",
+    "BowserCastleKey": "Bowsers Castle",
+    "CrystalPalaceKey": "Crystal Palace",
+    "KoopaFortressKey": "Koopa Bros Fortress",
+    "RedKey": "Crystal Palace",
+    "RuinsKey": "Dry Dry Ruins",
+    "TubbaCastleKey": "Tubbas Castle"
+}
+
+export const DUNEGON_KEYS = new Set(Object.keys(KEY_TO_DUNGEON));
+
+export const PARTNERS = new Set([
+    "Goombario",
+    "Kooper",
+    "Bombette",
+    "Parakarry",
+    "Bow",
+    "Watt",
+    "Sushie",
+    "Lakilester"
+]);
+
+export const SUPER_BLOCK_LOCATIONS = new Set([
+    "Elevator Attic Room (B2) - In SuperBlock",
+    "Metal Block Room (B3) - In SuperBlock",
+    "Blue Pushblock Room (B2) - In SuperBlock",
+    "Frozen Room (B3) - In SuperBlock",
+    "Under the Toad Town Pond - In SuperBlock",
+    "Train Station - In SuperBlock",
+    "S2E3 Oasis - In SuperBlock",
+    "Vertical Shaft - In SuperBlock",
+    "Stairs to Basement - In SuperBlock",
+    "RED Moving Platforms - In SuperBlock",
+    "SW Jungle (Super Block) - In SuperBlock",
+    "Fire Bar Bridge - In SuperBlock",
+    "Zipline Cavern - In SuperBlock",
+    "(NE) Elevators - In SuperBlock",
+    "(SE) Briar Platforming - In SuperBlock",
+    "Shiver Mountain Hills - In SuperBlock",
+    "Elevator Attic Room (B2) - In SuperBlock",
+    "Blue Pushblock Room (B2) - In SuperBlock",
+    "Metal Block Room (B3) - In SuperBlock",
+    "Frozen Room (B3) - In SuperBlock",
+    "Under the Toad Town Pond - In SuperBlock",
+    "Train Station - In SuperBlock",
+    "S2E3 Oasis - In SuperBlock",
+    "Vertical Shaft - In SuperBlock",
+    "Stairs to Basement - In SuperBlock",
+    "RED Moving Platforms - In SuperBlock",
+    "SW Jungle (Super Block) - In SuperBlock",
+    "Fire Bar Bridge - In SuperBlock",
+    "Zipline Cavern - In SuperBlock",
+    "(SE) Briar Platforming - In SuperBlock",
+    "(NE) Elevators - In SuperBlock",
+    "Shiver Mountain Hills - In SuperBlock"
+]);
+
+export const PROGRESSIVE_BADGES = new Set([
+    "ProgressiveJumpCharge",
+    "ProgressivePowerJump",
+    "ProgressivePowerSmash",
+    "ProgressiveQuakeHammer",
+    "ProgressiveSmashCharge"
+])
+
+export const GEAR_LOCATIONS: Set<string> = new Set([
+    "Super Boots Room - In Big Chest",
+    "Ultra Boots Room (B3) - In Big Chest",
+    "Jr. Troopas Playground - In Hammer Bush",
+    "Super Hammer Room - In Big Chest",
+    "Ultra Hammer Room - In Big Chest"
+]);
+
+export const LETTER_CHAIN_CHECKS: Set<string> = new Set([
+    "Goomba Village - Goompapa Letter Reward 1",
+    "Ruined Castle Grounds - Muss T. Letter Reward",
+    "Koopa Village 1 - Koover Letter Reward 1",
+    "Port District - Fishmael Letter Reward",
+    "Koopa Village 1 - Koover Letter Reward 2",
+    "Outpost 2 - Mr. E. Letter Reward",
+    "Gate District - Miss T. Letter Reward",
+    "Outpost 1 - Little Mouser Letter Reward",
+    "Foyer - Franky Letter Reward",
+    "Station District - Dane T. Letter Reward 1",
+    "Village Buildings - Red Yoshi Kid Letter Reward",
+    "Station District - Dane T. Letter Reward 2",
+    "Starborn Valley - Frost T. Letter Reward"
+]);
